refactor: create an Abstract Dmas
