+++ jakta-distributed/jakta-distributed-adapter/src/main/kotlin/it/unibo/jakta/agents/distributed/dmas/impl/DMasWithEmbeddedBroker.kt
@@ -1,20 +1,10 @@
 package it.unibo.jakta.agents.distributed.dmas.impl
 
 import it.unibo.jakta.agents.bdi.Agent
-import it.unibo.jakta.agents.bdi.actions.effects.AddData
-import it.unibo.jakta.agents.bdi.actions.effects.BroadcastMessage
-import it.unibo.jakta.agents.bdi.actions.effects.EnvironmentChange
-import it.unibo.jakta.agents.bdi.actions.effects.PopMessage
-import it.unibo.jakta.agents.bdi.actions.effects.RemoveAgent
-import it.unibo.jakta.agents.bdi.actions.effects.RemoveData
-import it.unibo.jakta.agents.bdi.actions.effects.SendMessage
-import it.unibo.jakta.agents.bdi.actions.effects.SpawnAgent
-import it.unibo.jakta.agents.bdi.actions.effects.UpdateData
 import it.unibo.jakta.agents.bdi.environment.Environment
 import it.unibo.jakta.agents.bdi.executionstrategies.ExecutionStrategy
 import it.unibo.jakta.agents.distributed.RemoteService
 import it.unibo.jakta.agents.distributed.broker.embedded.EmbeddedBroker
-import it.unibo.jakta.agents.distributed.dmas.DMas
 import it.unibo.jakta.agents.distributed.network.Network
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
@@ -32,7 +22,14 @@ internal class DMasWithEmbeddedBroker(
     override val network: Network,
     override val services: Iterable<RemoteService>,
     private val broker: EmbeddedBroker,
-) : DMas {
+) : AbstractDMas(
+    executionStrategy,
+    environment,
+    agents,
+    services,
+    network,
+) {
+
     init {
         agents.forEach { environment = environment.addAgent(it) }
     }
@@ -53,51 +50,6 @@ internal class DMasWithEmbeddedBroker(
                 network.subscribe(RemoteService("broadcast"))
             }
             executionStrategy.dispatch(self, debugEnabled)
-        }
-    }
-
-    override fun applyEnvironmentEffects(effects: Iterable<EnvironmentChange>) {
-        val externalEffects = network.getMessagesAsEnvironmentChanges()
-        (effects + externalEffects).forEach { environmentChange ->
-            when (environmentChange) {
-                is BroadcastMessage -> {
-                    runBlocking {
-                        launch(Dispatchers.Default) {
-                            network.broadcast(environmentChange)
-                        }
-                    }
-                    environment = environment.broadcastMessage(environmentChange.message)
-                }
-
-                is RemoveAgent -> {
-                    agents = agents.filter { agent -> agent.name != environmentChange.agentName }
-                    executionStrategy.removeAgent(environmentChange.agentName)
-                    environment = environment.removeAgent(environmentChange.agentName)
-                }
-
-                is SendMessage -> {
-                    if (services.map { it.serviceName }.contains(environmentChange.recipient)) {
-                        runBlocking {
-                            launch(Dispatchers.Default) {
-                                network.send(environmentChange)
-                            }
-                        }
-                    } else {
-                        environment = environment.submitMessage(environmentChange.recipient, environmentChange.message)
-                    }
-                }
-
-                is SpawnAgent -> {
-                    agents += environmentChange.agent
-                    executionStrategy.spawnAgent(environmentChange.agent)
-                    environment = environment.addAgent(environmentChange.agent)
-                }
-
-                is AddData -> environment = environment.addData(environmentChange.key, environmentChange.value)
-                is RemoveData -> environment = environment.removeData(environmentChange.key)
-                is UpdateData -> environment = environment.updateData(environmentChange.newData)
-                is PopMessage -> environment = environment.popMessage(environmentChange.agentName)
-            }
         }
     }
